made protagonist movement without collision

+++ Java_Gaming/src/main/model/Position.java
@@ -2,8 +2,8 @@ package main.model;
 
 public class Position {
 	private double x, y;
-	public void setX(int x) {this.x = x;}
-	public void setY(int y) {this.y = y;}
+	public void setX(double x) {this.x = x;}
+	public void setY(double y) {this.y = y;}
 	public double getX() {return x;}
 	public double getY() {return y;}
 	public Position(double x, double y) {
